Fix Soal Delete

+++ app/Modules/Admin/Controllers/SoalController.php
@@ -133,7 +133,9 @@ class SoalController extends Controller
         if ($soal->tes) {
           $soal->tes()->forceDelete();
         }
-        $soal->jadwal()->forceDelete();
+        if ($soal->jadwal) {
+          $soal->jadwal()->forceDelete();
+        }
         if ($soal->delete()) {
           return redirect()->back()->with('message', 'Data berhasil dihapus');
         }
